refactor(signing): drop underscore prefix from internal module names
Renames the 11 underscore-prefixed modules under src/adcp/signing/ to plain
names (e.g. _canonical.py -> canonical.py). The _prefix convention doesn't
match sibling packages (adcp/server, adcp/protocols, adcp/utils all use plain
names), and the signing tests already import these modules directly - if a
test needs it, the module isn't private.

Closes #189.
